move formatNumber to numbers.ts
